backend/pyproject.toml
[project]
name = "task-management"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Diogenes Fernandes", email = "diofeher@gmail.com" }
]

dependencies = [
    "fastapi[standard]>=0.115.4",
    "hypercorn>=0.16.0",
    "email-validator>=2.1.1",
    "pydantic>=2.7.1",
    "python-multipart>=0.0.9",
    "sqlmodel>=0.0.22",
    "pytest>=8.3.3",
    "ruff>=0.7.1",
    "passlib>=1.7.4",
    "pyjwt[crypto]>=2.9.0",
    "psycopg2>=2.9.10",
    "sqlalchemy-utils>=0.41.2",
    "pydantic-settings>=2.6.0",
    "mypy>=1.13.0",
    "pip>=24.3.1",
]
readme = "README.md"
requires-python = ">= 3.8"

# [project.scripts]
# dev = "rye run fastapi dev src"
# build = "docker build ."

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "mypy>=1.10.0",
    "isort>=5.13.2",
    "black>=24.4.2",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/task_management"]

[tool.ruff]
# Set the maximum line length to 79.
line-length = 79

[tool.mypy]
ignore_missing_imports=true
